Completed version 2
Created areas for users whether signed up or not where they can view
all of the shows and albums that users have added in their personal
pages. Added the ability for users to add albums to their pages. This
allows them to upload full images for anyone to see so that the band
may showcase their art. And lastly, I created a fully functional
registration form that only requires users to enter a band name,
password, and email.